Test AXTree algorithms by generating all possible trees of size 4.
Tests updating from every possible tree with ids 1...4 to every other
tree with ids 1...4. It's possible to run on larger tree sizes but that
takes too long to run on the bots.

This test helped catch several corner cases that didn't work previously.
(It's not clear if these cases would be possible in Blink now, but this
helps ensure that even future Blink bugs can't cause an invalid AXTree or
a crash.)

BUG=316726
R=dtseng@chromium.org

Review URL: https://codereview.chromium.org/90873002