Omnibox: Launch Stop Timer Behavior
Previously, we ran an experiment that caused all autocomplete providers
to stop delivering results after 1.5 seconds.  This was caused by the
assumption that results coming in after that length of time are
disruptive, likely causing more harm than benefit.  We decided to launch
this experiment.  This change turns on that behavior for everyone.

This change also converts the old "use stop timer" field trial code to
a parameterized form, so we can run future experiments with shorter
timeouts.

BUG=

Review URL: https://codereview.chromium.org/117233009