When stopping QTKit video capture, call stopRunning before removeInput:.
Empirically, this prevents a browser hang when the active camera is
unplugged.  The unplug event triggers both stopping the capturer and
enumerating the available input devices at the same time.  It seems
that calling removeInput: on a running capturer after an error
produces a deadlock, but stopping the capturer first resolves the
problem.

BUG=285053

Review URL: https://codereview.chromium.org/120963003