mac: Fix a crash in the history swiper for magic mouse momentum events.
The 'trackSwipeWithEvent:' api throws an exception if the event passed in is a
momentum scroll event. We started using momentum scroll events in our logic for
magic trackpads, since we're working directly with hardware touches. We failed
to filter out momentum events when falling back to the 'trackSwipeWithEvent:'
api for magic mouse events.

BUG=331506

Review URL: https://codereview.chromium.org/124313002