Refactor the component updater update check code into its own class.
The idea here is to decouple the code that does update checks and
speaks the Google Update protocol from the rest of the service.

If we want to secure the update channel using our custom message signing,
then changes will be localized to the UpdateChecker module and not touch
the service code.

Also, this change avoids spilling protocol details and XML formatting in
the service module by encapsulating them in the UpdateChecker class.

BUG=330956

Review URL: https://codereview.chromium.org/119843003