Files.app: Re-implement the notifications for removable devices in JavaScript.
Currently the notifications for removable devices are implemented in C++. But
now we have the extension notification API, so we can implement thme in
JavaScript.

This CL adds an event of device state change to the file manager private API,
and handles it to show the notifications.

BUG=330096
TEST=insert USB on link

Review URL: https://codereview.chromium.org/113893008