Pepper: Split PpapiHostMsg_ChannelCreated use.
PpapiHostMsg_ChannelCreated is used for two purposes:
  * Notifying the browser that a browser<->plugin channel has been created, and
  * Notifying that browser that a renderer<->plugin channel has been created.

In this case renderer<->plugin channel is for NaCl-based plugins, and is
special because it does some inspection of message contents and passes on the
message to the untrusted side of the plugin.

I've changed this because I think it makes the code easier to read, and I'm
adding an additional renderer<->plugin channel anyway.

BUG=
TBR=mseaborn,jschuh

Review URL: https://codereview.chromium.org/98183012