GTTF: Include info about disabled tests and platforms in JSON test summary.
This will make it possible to analyze which tests are disabled on specific
platforms, and whether the failures are really platform-specific.

Including list of all tests in the summary makes it possible to e.g. detect
when a test first appeared (or disappeared), which platforms a test is compiled
on, or whether some tests have been filtered out.

BUG=236893
R=sky@chromium.org

Review URL: https://codereview.chromium.org/93603008