Run the incognito title-changing content_script extension for tests at document_end
In Blink r164582 ("Don't set document.title when there is no <head>") Blink
aligns the behavior of the "document.title" property with the HTML specification,
and now disregards users who set the title when no <head> or <title> element
exist yet.

This content_script extension runs at document_start, at which time no DOM is
available yet. Changing this to document_end means that the DOM is available, and
it's thus able to change the page's title (the sole purpose of the extension).

BUG=332106
R=wjmaclean@chromium.org

Review URL: https://codereview.chromium.org/117993006