Mirror images for RTL languages at build time.
This adds a build step to generate mirrored images for use in
right-to-left (RTL) languages. Images are mirrored by flipping the
original image over the vertical axis. Every image must be explicitly
listed as mirrorable or non-mirrorable in a config file.

The goal: ensure that our RTL image assets are always up-to-date.

BUG=290225
NOTRY=true

Review URL: https://codereview.chromium.org/106173002