mac: Add a missing animation context to bookmark bar controller.
The animation was not specifying an NSAnimationContext or duration, which meant
that the duration was not well defined. I used the same value that I observed
on my machine prior to my change. It seems a bit slow, but presumably it's the
value most users are seeing.

BUG=NONE
TEST=NONE

Review URL: https://codereview.chromium.org/110753004