Do not dispatch events to inactive ephemeral apps
With the exception of the chrome.app.runtime.onLaunched() event,
events will only be dispatched to ephemeral apps that are running.
This was implemented by adding a "can_load_ephemeral_apps" flag to
extensions::Event. Only events which enable this flag will have the
ability to load the event page of apps.

BUG=312460
TEST=browser_tests (EphemeralAppBrowserTest.*)

Review URL: https://codereview.chromium.org/118033002