base: Discardable memory types.
This adds a DiscardableMemoryType enum and a function that
can be used to select what type of discardable memory to
use. This is generally useful for debugging purposes and
when evaluating the performance of one implementation vs
another.

The unit test framework for discardable memory is improved
slightly by testing every supported type on each platform
rather then just the one specific type.

Furthermore, it allows us to expose ashmem based
discardable memory on ChromeOS in about:flags before we
make it default.

No change in behavior unless the new
--use-discardable-memory=type switch is used.

BUG=327516
TEST=base_unitttest --gtest_filter=DiscardableMemory*
TBR=tomhudson@google.com

Review URL: https://codereview.chromium.org/114923005