Fallback to on-line password check if local fails.
Relying on local-auth is not sufficient because locking can be enabled when no local credentials are present.  Also, changing the password outside of this instance of Chrome would require using the old password to unlock, now no longer the case.

This affects only new-profile-management.

BUG=328017

Review URL: https://codereview.chromium.org/99373008