Use apple-touch-icons for streamlined hosted apps.
This patch allows streamlined hosted apps to use the apple touch icons
for generating hosted apps.

To use the icons, it is necessary to size them to
extension_misc::kExtensionIconSizes to make the manifest handler agree
to use the icons.

Added TabHelper::ConstrainBitmapsToSizes which finds the
closest, not-smaller bitmap for each size and downsizes it. At most each
bitmap is used for 1 size in order to prevent generating resized icons
which would be stored on disk.

BUG=318607

Review URL: https://codereview.chromium.org/105873009