OpaqueBrowserFrameViewPlatformSpecific::Create no longer returns NULL.
This avoids duplicating the default logic for all of its methods.
Previously, every time a method was called, we had to test for NULL and
assume the default value for that method (even though the default method
implementation already exists; it was never used). Now the caller
assumes platform_observer_ is non-NULL (via a DCHECK) and it is the base
class's responsibility to provide default methods.

Review URL: https://codereview.chromium.org/124893002