Add error handling for supervised user import flow.
Add a check if the user tries to create a profile with the
name of an already existing supervised user. In that case,
the user is not allowed to create a new such supervised user
profile, instead he can directly import the existing supervised
user.
A screenshot can be found in the bug description.

BUG=282464
TEST=browser_tests

Review URL: https://codereview.chromium.org/125993002