Files.app: Adjust tab index for the save file diaog.
In the save file dialog, the file name input has a initial focus.  Previously,
the file name input has a tabindex=10, and when the focus goes back to the
document, it moves to the file name input again because of the refocus method of
Filemanager class.

This CL change the tabindex of elements on the save-as dialog's footer and allow
users to explorer all controls in Files.app.

BUG=332051
TEST=manually

Review URL: https://codereview.chromium.org/127843002