Prepare for including compositing reasons in GraphicsLayerDebugInfo
Currently, the debug info is assumed to include only layout rects.
This is a bummer. This CL allows for blink to return an arbitrary JSON
object that will simply be merged into the base::DictionaryValue
we populate for about::tracing. This will allow us to return
arbitrary debug info for tracing from blink.

BUG=None

Review URL: https://codereview.chromium.org/128303003