Create signin component and componentize TokenWebData.
This CL creates the signin layered component. It additionally componentizes
TokenWebData and moves it into the signin component. The refactorings performed
on TokenWebData are identical to ones performed on AutofillWebDataService:
- threads are injected rather than being obtained from content::BrowserThread
- //chrome code gets the TokenWebData for a Profile via WebDataServiceFactory
instead of TokenWebData::FromBrowserContext.

This CL is an unchanged reland of https://codereview.chromium.org/109783003/,
which had been speculatively reverted to see if it was causing ChromeOS build
failures. The speculative revert unfortunately stuck as it happened to land in
the same run as the CL that actually fixed the (unrelated) ChromeOS build
failures: https://codereview.chromium.org/129343003/.

BUG=332107
TBR=noamsml

Review URL: https://codereview.chromium.org/130813003