Show connection type in web-app.
This is mostly boilerplate to get the OnSessionRouteChange notification up to the web-app. The web-app then includes it in the stats panel. In the common case where all channels have the same type, it displays it as "connection: <type>"; if the types differ they are displayed individually.

BUG=134528

Review URL: https://codereview.chromium.org/126123002