Implement registration job ordering
To handle the case of multiple registration requests coming in at the same time,
the ServiceWorkerJobCoordinator now keeps a queue of registration events for each pattern.

When a new registration request comes in for a pattern, it looks to see if the most recent request (at the back of the queue) is identical, and if so, piggy-backs on that request rather than creating a new one.

BUG=285976

Review URL: https://codereview.chromium.org/126603002