Added memory and V8 metrics to indexeddb_perf results
We are trying to deprecate the page_cycler.indexeddb performance test.
One of the things done by that test and not indexeddb_perf is the
measurement of V8 memory statistics. This change brings V8 stats
into indexeddb_perf - but only a subset of the full measurements.
Opted for a smaller subset as the full set has too much unrelated
stuff, and we really want to catch things like unreleased objects,
etc.

Review URL: https://codereview.chromium.org/123553003