Linux: fix getting termination status for known dead processes.
While a process is being destroyed by the kernel, it will loose
its IPC file descriptors which signals to the browser that the process is dead.

In that timeframe, doing a non blocking waitpid() is racy and
can indicate the process as still being running while the kernel is
still destroying it.

We unify Linux with OS X and use known_dead even in the non Zygote case.

BUG=332635
NOTRY=true

Review URL: https://codereview.chromium.org/130633002