Use web-based sign-in, rather than inline signin, for v1 and v2 apps.
The remoting browser test started failing for the v2 web-app, because
authentication now redirects by default to the inline signin flow.
To fix this, we enforce web-based sign-in by passing in the required flag.

Tested that existing MANUAL_Me2Me_Connect_Local_Host, which was passing for the v1 web-app and failing for the v2 web-app, now passes for both with this change.

BUG=NONE

Review URL: https://codereview.chromium.org/132433005