Explicitly use Cmd+w for "Close Window" in apps. (Mac)
Previously, the "Close Window" item is copied from Chrome's "File" menu.
However, the key equivalent of this item changes between Cmd+Shift+w and
Cmd+w depending on whether a tabbed window has focus. This happens in
-[AppControllerMac adjustCloseWindowMenuItemKeyEquivalent:].

This causes a bug when the item in the app's menu does not match the one
in Chrome's. OSX will remove the keyboard shortcut for any items that have
a matching title but different key equivalents. 

In this CL, the key equivalent is always set to Cmd+w in the app's menu. The "Close Window" item in Chrome will be changed to Cmd+w when an app
window gains focus. Duplicate items are allowed to share a key equivalent
if their titles match.

BUG=302117

Review URL: https://codereview.chromium.org/130673002