[NaCl SDK] Change KernelProxy::GetCWD() to allocate path.
When supplying a zero size to getcwd() typical behavior is to allocate
internally and return a pointer.  This behavior was already implemented, but
disabled by checking if size <= 0 and erroring with EINVAL.  As size_t is
positive definite, this entire check can be removed.

R=noelallen@chromium.org, sbc@chromium.org, binji@chromium.org
BUG=

Review URL: https://codereview.chromium.org/122943005