base: Fix registering of memory pressure listeners used when emulating discardable memory.
Memory pressure listeners need to be created on a thread with a message
loop current. Discardable memory is often used on worker threads that
don't have a message loop and with the current system where listeners
are registered on first use, the result is that we're most likely never
notified of memory pressure.

This adds a function to the discardable memory interface that allows us
to register memory pressure listeners on a thread, and at a time, when
we know a message loop is current.

BUG=332570
TBR=jamesr@chromium.org

Review URL: https://codereview.chromium.org/129963002