app_shell: Fix PrefsTabHelper dependency of ExtensionHost
The default preference values are fine for app_shell's pages (as they are
for content_shell, which doesn't use PrefsTabHelper).  Delegate out the
creation of PrefsTabHelper so we only create them in Chrome.

With this patch app_shell can launch a renderer before crashing.

BUG=332435
TEST=browser_tests, manual testing that extensions still work

Review URL: https://codereview.chromium.org/132993002