Force download_from_google_storage --config to produce a read-only token
When download_from_google_storage --config is run, it should be implied that
the user just wants to download, not upload.  This change passes the '-r' flag
into gsutil config, which requests a read-only scoped token rather than write
token.  This is saved in ~/.boto.depot_tools so that it doesn't conflict
with a ~/.boto file crated later that may have write permissions.


BUG=

Review URL: https://codereview.chromium.org/135153002