Fix for testing existence of oobe in login code.
We were using GetOobe to figure out if the oobe page existed. This creates an InspectorBackend, which creates an InspectorPage, which calls _EnablePageNotifications, which calls SyncRequest with a timeout of 60 sec. If the page goes away right before SyncRequest is called, we could end up waiting 60 sec.

In general, GetOobe is too expensive if all we're interested in is checking the existence of oobe, so fix all instances of such checks to use OobeExists instead.

BUG=331940
TEST=manual
NOTRY=True

Review URL: https://codereview.chromium.org/134043005