Tweak ifdefs for MessageLoopForUI::Add/RemoveObserver
This is necessitated by our building of base for tools. In particular
if I set OS_ANDROID and USE_AURA then the host is built with OS_LINUX
and USE_AURA, but none of the other flags we normally set if USE_AURA
is set for OS_LINUX.

And yes, I really need to spend time and refactor messageloop. It is
such a mess of ifdefs right now.

BUG=none
TEST=none
R=darin@chromium.org

Review URL: https://codereview.chromium.org/129873003