Fix apps page footer shrinking/overlap for small window sizes.
This CL shrinks the logo instead of hiding the nav dots and makes the
nav dots shrink to the same size by using the new*** flexbox spec. It
also delays dot layout until the footer logo is loaded as this is
necessary for accurate measurements.

BUG=332350

Review URL: https://codereview.chromium.org/128013002