chromeos: Do not hide the cursor if a mouse device is used to generate touch events.
A mouse device can be used to generate touch events for development purposes on
a workstation (using --touch-devices flag). However, the default behaviour in
chromeos of hiding the cursor on touch-press makes it somewhat difficult to use
properly. So avoid hiding the cursor in such circumstances.

BUG=none
R=tdanderson@chromium.org

Review URL: https://codereview.chromium.org/130433004