chromium-status: fix interpretation of week days when localizing timestamps
The short timestamps the server outputs does not include a year:
	Fri, 10 Jan 18:11
this means when we try to localize things, javascript picks a year
for us (2001) and uses that to figure out the new day of the week.

For now, assume all short timestamps represent this year.  It'll be right
the majority of the time (only when showing Dec dates during Jan will it
be slightly off).

BUG=chromium:314501
TEST=ran an appengine and checked the day of the week to make sure it was correct

Review URL: https://codereview.chromium.org/134403003