Port blink_perf animation tests over as tough_animation_cases
Currently the blink_perf animation tests rely on Javascript to measure
frame rates. This is not ideal and can give very inaccurate results for
animations that run on the compositor.
This patch copies the animation perf tests to src/ as smoothness tests
for more accurate reading and more detailed analysis.

BUG=305509

Review URL: https://codereview.chromium.org/105033004