Files.app: Blur the focus when window the state is changed.
Previously the focus is being on the gear menu even when the window state is
changed. It causes the situation where the two gear and maximized buttons look
focused. This CL add a call for the blur method when the window state is
changed.

BUG=333196
TEST=manually

Review URL: https://codereview.chromium.org/130563007