Fix updating network lists after removal of policy from a network.
If a visible network was removed from policy, the UI didn't update correctly and continued to show the old network as managed.

The problem was in the favorite list not correctly updating.
Additionally, there was a potential subtle bug in PolicyApplicator, if the SSID of a Wifi network is changed in policy. Now, it's ensured that at first all profile entries are removed and then the new configuration is written to Shill.

BUG=278101

Review URL: https://codereview.chromium.org/107733003