Revert patch to invoke cursor-hiding logic on touch for Windows
This is a revert of r240976, which updates our internal tracking
of the cursor state (visibility and enabledness) on touch for
Windows. This change is responsible for a handful of regressions
(see crbug.com/333952 for a list) and has already been reverted
from the M33/1750 branch (see crbug.com/332477).

Re-introducing this code will require that crbug.com/332430
be resolved first.

BUG=333952
R=sky@chromium.org

Review URL: https://codereview.chromium.org/134643009