Don't set the scroll styles (WS_VSCROLL and WS_HSCROLL) for WS_POPUP windows.
This causes issues with select boxes on Windows 7 where hovering at the end of the window returns the scroll WM_NCHITTEST
codes. Works fine on Windows 8. In any case we don't want the scrolling styles set on windows other than the main Chrome window
which is the only window which should be receive mousewheel messages.

I moved the scroll style setting code from the HWNDMessageHandler::OnCreate function to HWNDMessageHandler::Init function as that
would prevent the initial WM_SIZE message from hiding the scrollbar. The other change is to hide the scrollbars and readd the
scroll styles if we are sizing the window, when the sizing completes. Basically when we receive the WM_EXITSIZEMOVE message.
For normal sizing operations we continue to do this in WM_SIZE as before.

From testing on my thinkpad with Win7, desktop with Win7 and Win8 this works well.

BUG=334454,334541
R=sky@chromium.org, sky

Review URL: https://codereview.chromium.org/137403008