Wait for SHOW_PRESS gesture to avoid flake in root_window_unittest.cc
Previously in RootWindowTest.TouchMovesHeld, sometimes a SHOW_PRESS
gesture would be fired, and sometimes it wouldn't, depending on
whether or not the 5ms timer used in tests for the SHOW_PRESS
gesture was fired. Now we explicitly wait for the gesture to be
fired before continuing.

Marked flaky here: https://codereview.chromium.org/136013002/

TEST=RootWindowTest.TouchMovesHeld
BUG=333644

Review URL: https://codereview.chromium.org/136843003