Avoid loading the last used browser profile in app_controller_mac when it's not needed.
Currently -[AppController applicationDidFinishLaunching:] and
-[AppController windowLayeringDidChange:] can trigger a synchronous load
of the last used _browser_ profile, even when no browsers are being
opened, and that profile is not required. This causes delays when the
app launcher or app shims are launched for a profile that doesn't match
the last used browser profile, because both profiles must be loaded and
initialized.

This change moves ChromeBrowserMain's GetStartupProfilePath() from an
anonymous namespace to startup_browser_creator.h so that AppController
can share the logic that determines the profile loaded when the browser
process starts up.

The shim process is tweaked slightly to ensure it passes
`--show-app-list` for the app launcher shim, similar to the Windows app
launcher. This allows a newly created browser process to determine that
the first thing it will be doing is showing a launcher.

BUG=178260, 262848

Review URL: https://codereview.chromium.org/127343004