Fix an issue that MediaController is disaled when entering fullscreen
In the old contentVideoView.java, we create the MediaController first in the openVideo() call.
Then we call updateMediaMetadata() to enable to media controller and set the metadata.
With ContentVideoViewLegacy.java, the ordering between updateMediaMetadata() and mediaController creation is reversed. That causes the controller to become disabled.
This change posts a task when RequestMediaMetadata() is called to solve the above issue, and to avoid the reentrance issue.

BUG=334635

Review URL: https://codereview.chromium.org/141243002