Reenable linux_aura on the main builders.
We fixed the browser test slowdown that lead to linux_aura being
disabled in r235280. (Different test suites are faster or slower, but
overall, tests run faster.)

We are turning this on primarily to get perf data; if things look
acceptable, we hope to cut a dev channel this week.

BUG=319075,125106,303342,318961
NOTRY=True

Review URL: https://codereview.chromium.org/136813004