Play automatically when reloading the audio player with new songs.
Because of introducing an async call during initialization, the new appstate was being changed. Playing a song adds a 'time' field to the appstate. Having a time field causes pausing the audio player once the initialization is finished.

To avoid this, a copy is made, to avoid altering the input 'playlist' object.

TEST=Tested manually. Partly browser tests.
BUG=335422

Review URL: https://codereview.chromium.org/141273004