Back out trunk r245361.
Go back to KeystoneRegistration.framework 1.1.0.3659. The new registration
framework 1.2.0.6782, when using -[KSRegistration
checkForUpdateWasUserInitiated:YES], is not compatible with an older installed
Keystone version such as the current 1.1.0.3973. This results in the About
page showing "Update failed (error: 11)" and this message being printed to
stderr:

ksadmin: unrecognized option `--user-initiated'

In this case, updating is not broken, and Keystone will continue to update
Chrome in the background, but no indication of update availability will be
made in Chrome’s about page.

> Update KeystoneRegistration.framework to 1.2.0.6782.
> 
> The big change in this version is that an installation attempt initiated by
> the application can show up as ondemand on the server. This will allow us to
> provide throttled Mac downloads (such as a "5% push") with correct behavior.
> 
> BUG=35495
> R=thakis@chromium.org
> 
> Review URL: https://codereview.chromium.org/141193002

BUG=35495
TBR=mark@chromium.org

Review URL: https://codereview.chromium.org/141583005