Disable TabCaptureApiTest.EndToEnd in debug or on OSMesa.
This test is quite slow, and gets slower under a real compositor that
tests the code paths we ship. After we use a real compositor in
browser tests (https://codereview.chromium.org/120313002/) this test
takes 18seconds on a release linux_chromeos build in a z620 machine
and ends up timing out on bots.

So, disable the test on debug builds, and in release builds don't run
the test when using OSMesa, since this test requires pixel results
it requires OSMesa output which is very slow. It should still run on
the ChromeOS VMs and Mac/Android where we do not use OSMesa.

R=miu, piman
BUG=270918

Review URL: https://codereview.chromium.org/141953010