Switch off always-on-top bit when the app window overlaps the taskbar
To prevent apps from covering and spoofing the Windows taskbar, the
always-on-top bit will be silently switched off when there is any
overlap between the app window and the Windows taskbar. The bit is
reinstated when the window moves away from the taskbar.

BUG=320107
TEST=Use the Window State sample Chrome app to test manually.
On Win 7, verify that when an app window with always-on-top enabled
cannot be above the taskbar. See bug (comment #30) for known issues.

Review URL: https://codereview.chromium.org/137533008