Disable the idle handler timer when no action is expected.
On Android, the webkit timer are disabled after the application is in
the background for 5 minutes. When this happens, there is no reason to
let the idle handler timer run indefinitly. This CL disables the timer
once v8 reported that it will not take any further action.

BUG=333875

Review URL: https://codereview.chromium.org/136903002