Limit display of the virtual keyboard to state changes triggered from a user gesture.
This patch reuses OnTextInputStateChanged, which was previously Android specific, but which includes not only type information, but whether the virtual keyboard should be displayed. This is the first step in consolidating the IPC messages between RenderWidget and RenderWidgetHostView for IME messages.  Ideally, we can phase out OnTextInputTypeChanged in favor of the approach used by Android across all platforms.



BUG=289659, 294191, 331690

Review URL: https://codereview.chromium.org/29943002