apps: Install the shaped window-targeter on the app windows.
Previously in chromium, a ShapedNativeAppWindowTargeter was introduced in
r245757 to make sure hit-testing for events work correctly for windows with
custom shapes. However, the targeter was never actually installed on any
window, causing it to not function properly. In this episode, we install
the shaped targeter on the window, and add a test (with some relevant
refactoring) to make sure it functions correctly.

BUG=318879
R=benwells@chromium.org

Review URL: https://codereview.chromium.org/134733004