Update net/third_party/nss to NSS 3.15.5 Beta 2.
alpn.patch and paddingextension.patch have been committed in the NSS
upstream.

In addition to calling SSL_SetNextProtoNego, the SSL_ENABLE_NPN and
SSL_ENABLE_ALPN options need to be used to control whether each of
NPN and APLN will be negotiated. SSL_ENABLE_NPN is enabled by default,
but we enable it explicitly to be safe.

R=agl@chromium.org
BUG=331625
TEST=none

Review URL: https://codereview.chromium.org/142283002