Rename the UseRealGLContexts() to EnablePixelOutput().
Now we always use a real gl context, this function name is wrong. Also
add a command line flag to force pixel output to be enabled for visual
debugging, as --enable-pixel-output-in-tests.

A number of tests set UseRealGLContexts() for reasons that don't exist
anymore so clean these up to try only calling EnablePixelOutput() from
tests that actually compare the output pixels as a criteria for passing.

R=piman, sky
BUG=270918

Review URL: https://codereview.chromium.org/143893008